requirements.txt
docopt
future
imgaug
matplotlib
opencv-python
pandas
pillow
psutil
scikit-image
scikit-learn
scipy
tensorboard
tensorboardx
termcolor
tqdm
numpy
ruamel.yaml
pyyaml
django-tailwind
django_browser_reload